+++ src/cache/BlockIndex.h
@@ -56,7 +56,7 @@ namespace cache {
 			uint64_t transactionNr, 
 			int32_t fileCursor, 
 			const uint32_t* addressIndices,
-			uint8_t addressIndiceCount
+			uint16_t addressIndiceCount
 		);
 
 		//! \brief add transactionNr - fileCursor pair to map if not already exist
